requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncio==3.4.3
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
opensearch-py==2.2.0
pyaes==1.6.1
pyasn1==0.5.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
Telethon==1.28.5
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
